requirements.txt

Pip Package             Version
------------------- ----------

docutils==0.17.1
Kivy==2.0.0
kivymd==0.104.2
Kivy-Garden==0.1.4
kivy_deps.sdl2==0.4.2
kivy_deps.glew==0.3.0
matplotlib==3.2.1
numpy==1.21.2
pandas==1.0.3
Pillow==8.3.2 
python-dateutil==2.8.2
pygments==2.10.0
pypiwin32==223
regex==2021.7.6
wordcloud==1.8.1
virtualenv==20.0.12
